requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
python-dotenv==1.0.0
drf-spectacular==0.27.0